application: moved admin http server to separate class
Factored out the admin API http server to into separate class.
