+++ src/v/redpanda/application.cc
@@ -34,15 +34,9 @@
 #include "pandaproxy/proxy.h"
 #include "platform/stop_signal.h"
 #include "raft/service.h"
-#include "redpanda/admin/api-doc/config.json.h"
-#include "redpanda/admin/api-doc/kafka.json.h"
-#include "redpanda/admin/api-doc/partition.json.h"
-#include "redpanda/admin/api-doc/raft.json.h"
-#include "redpanda/admin/api-doc/security.json.h"
+#include "redpanda/admin_server.h"
 #include "resource_mgmt/io_priority.h"
 #include "rpc/simple_protocol.h"
-#include "security/scram_algorithm.h"
-#include "security/scram_authenticator.h"
 #include "storage/chunk_cache.h"
 #include "storage/directories.h"
 #include "syschecks/syschecks.h"
@@ -55,18 +49,10 @@
 #include <seastar/core/prometheus.hh>
 #include <seastar/core/smp.hh>
 #include <seastar/core/thread.hh>
-#include <seastar/http/api_docs.hh>
-#include <seastar/http/exception.hh>
-#include <seastar/http/file_handler.hh>
 #include <seastar/json/json_elements.hh>
 #include <seastar/net/tls.hh>
 #include <seastar/util/defer.hh>
 
-#include <boost/algorithm/string/classification.hpp>
-#include <boost/algorithm/string/split.hpp>
-#include <rapidjson/document.h>
-#include <rapidjson/stringbuffer.h>
-#include <rapidjson/writer.h>
 #include <sys/utsname.h>
 
 #include <chrono>
@@ -108,8 +94,8 @@ int application::run(int ac, char** av) {
                 initialize();
                 check_environment();
                 setup_metrics();
-                configure_admin_server();
                 wire_up_services();
+                configure_admin_server();
                 start();
                 app_signal.wait().get();
                 vlog(_log.info, "Stopping...");
@@ -246,25 +232,17 @@ void application::check_environment() {
     }
 }
 
-/**
- * Prepend a / to the path component. This handles the case where path is an
- * empty string (e.g. url/) or when the path omits the root file path directory
- * (e.g. url/index.html vs url//index.html). The directory handler in seastar is
- * opininated and not very forgiving here so we help it a bit.
- */
-class dashboard_handler final : public ss::httpd::directory_handler {
-public:
-    dashboard_handler()
-      : directory_handler(*config::shard_local_cfg().dashboard_dir()) {}
-
-    ss::future<std::unique_ptr<ss::httpd::reply>> handle(
-      const ss::sstring& path,
-      std::unique_ptr<ss::httpd::request> req,
-      std::unique_ptr<ss::httpd::reply> rep) override {
-        req->param.set("path", "/" + req->param.at("path"));
-        return directory_handler::handle(path, std::move(req), std::move(rep));
-    }
-};
+static admin_server_cfg
+admin_server_cfg_from_global_cfg(scheduling_groups& sgs) {
+    return admin_server_cfg{
+      .endpoint = config::shard_local_cfg().admin(),
+      .tls = config::shard_local_cfg().admin_api_tls(),
+      .dashboard_dir = config::shard_local_cfg().dashboard_dir(),
+      .admin_api_docs_dir = config::shard_local_cfg().admin_api_doc_dir(),
+      .enable_admin_api = config::shard_local_cfg().enable_admin_api(),
+      .sg = sgs.admin_sg(),
+    };
+}
 
 void application::configure_admin_server() {
     auto& conf = config::shard_local_cfg();
@@ -272,99 +250,13 @@ void application::configure_admin_server() {
         return;
     }
     syschecks::systemd_message("constructing http server").get();
-    construct_service(_admin, ss::sstring("admin")).get();
-    // configure admin API TLS
-    if (conf.admin_api_tls().is_enabled()) {
-        _admin
-          .invoke_on_all([this](ss::http_server& server) {
-              return config::shard_local_cfg()
-                .admin_api_tls()
-                .get_credentials_builder()
-                .then([this, &server](
-                        std::optional<ss::tls::credentials_builder> builder) {
-                    if (!builder) {
-                        return ss::now();
-                    }
-
-                    return builder
-                      ->build_reloadable_server_credentials(
-                        [this](
-                          const std::unordered_set<ss::sstring>& updated,
-                          const std::exception_ptr& eptr) {
-                            cluster::log_certificate_reload_event(
-                              _log, "API TLS", updated, eptr);
-                        })
-                      .then([&server](auto cred) {
-                          server.set_tls_credentials(std::move(cred));
-                      });
-                });
-          })
-          .get0();
-    }
-    if (conf.dashboard_dir()) {
-        _admin
-          .invoke_on_all([](ss::http_server& server) {
-              server._routes.add(
-                ss::httpd::operation_type::GET,
-                ss::httpd::url("/dashboard").remainder("path"),
-                new dashboard_handler());
-          })
-          .get0();
-    }
-    ss::prometheus::config metrics_conf;
-    metrics_conf.metric_help = "redpanda metrics";
-    metrics_conf.prefix = "vectorized";
-    ss::prometheus::add_prometheus_routes(_admin, metrics_conf).get();
-    if (conf.enable_admin_api()) {
-        syschecks::systemd_message(
-          "enabling admin HTTP api: {}", config::shard_local_cfg().admin())
-          .get();
-        auto rb = ss::make_shared<ss::api_registry_builder20>(
-          conf.admin_api_doc_dir(), "/v1");
-        _admin
-          .invoke_on_all([this, rb](ss::http_server& server) {
-              auto insert_comma = [](ss::output_stream<char>& os) {
-                  return os.write(",\n");
-              };
-              rb->set_api_doc(server._routes);
-              rb->register_api_file(server._routes, "header");
-              rb->register_api_file(server._routes, "config");
-              rb->register_function(server._routes, insert_comma);
-              rb->register_api_file(server._routes, "raft");
-              rb->register_function(server._routes, insert_comma);
-              rb->register_api_file(server._routes, "kafka");
-              rb->register_function(server._routes, insert_comma);
-              rb->register_api_file(server._routes, "partition");
-              rb->register_function(server._routes, insert_comma);
-              rb->register_api_file(server._routes, "security");
-              ss::httpd::config_json::get_config.set(
-                server._routes, []([[maybe_unused]] ss::const_req req) {
-                    rapidjson::StringBuffer buf;
-                    rapidjson::Writer<rapidjson::StringBuffer> writer(buf);
-                    config::shard_local_cfg().to_json(writer);
-                    return ss::json::json_return_type(buf.GetString());
-                });
-              admin_register_raft_routes(server);
-              admin_register_kafka_routes(server);
-              admin_register_security_routes(server);
-          })
-          .get();
-    }
-
-    with_scheduling_group(_scheduling_groups.admin_sg(), [this] {
-        return rpc::resolve_dns(config::shard_local_cfg().admin())
-          .then([this](ss::socket_address addr) mutable {
-              return _admin
-                .invoke_on_all<ss::future<> (ss::http_server::*)(
-                  ss::socket_address)>(&ss::http_server::listen, addr)
-                .handle_exception([this](auto ep) {
-                    _log.error("Exception on http admin server: {}", ep);
-                    return ss::make_exception_future<>(ep);
-                });
-          });
-    }).get();
-
-    vlog(_log.info, "Started HTTP admin service listening at {}", conf.admin());
+    construct_service(
+      _admin,
+      admin_server_cfg_from_global_cfg(_scheduling_groups),
+      std::ref(partition_manager),
+      controller.get(),
+      std::ref(shard_table))
+      .get();
 }
 
 static storage::kvstore_config kvstore_config_from_global_config() {
@@ -795,360 +687,7 @@ void application::start_redpanda() {
         _wasm_event_listener->start().get();
         pacemaker.invoke_on_all(&coproc::pacemaker::start).get();
     }
-}
-
-void application::admin_register_raft_routes(ss::http_server& server) {
-    ss::httpd::raft_json::raft_transfer_leadership.set(
-      server._routes, [this](std::unique_ptr<ss::httpd::request> req) {
-          raft::group_id group_id;
-          try {
-              group_id = raft::group_id(std::stoll(req->param["group_id"]));
-          } catch (...) {
-              throw ss::httpd::bad_param_exception(fmt::format(
-                "Raft group id must be an integer: {}",
-                req->param["group_id"]));
-          }
-
-          if (group_id() < 0) {
-              throw ss::httpd::bad_param_exception(
-                fmt::format("Invalid raft group id {}", group_id));
-          }
-
-          if (!shard_table.local().contains(group_id)) {
-              throw ss::httpd::not_found_exception(
-                fmt::format("Raft group {} not found", group_id));
-          }
-
-          std::optional<model::node_id> target;
-          if (auto node = req->get_query_param("target"); !node.empty()) {
-              try {
-                  target = model::node_id(std::stoi(node));
-              } catch (...) {
-                  throw ss::httpd::bad_param_exception(
-                    fmt::format("Target node id must be an integer: {}", node));
-              }
-              if (*target < 0) {
-                  throw ss::httpd::bad_param_exception(
-                    fmt::format("Invalid target node id {}", *target));
-              }
-          }
-
-          vlog(
-            _log.info,
-            "Leadership transfer request for raft group {} to node {}",
-            group_id,
-            target);
-
-          auto shard = shard_table.local().shard_for(group_id);
-
-          return partition_manager.invoke_on(
-            shard, [group_id, target](cluster::partition_manager& pm) mutable {
-                auto consensus = pm.consensus_for(group_id);
-                if (!consensus) {
-                    throw ss::httpd::not_found_exception();
-                }
-                return consensus->transfer_leadership(target).then(
-                  [](std::error_code err) {
-                      if (err) {
-                          throw ss::httpd::server_error_exception(fmt::format(
-                            "Leadership transfer failed: {}", err.message()));
-                      }
-                      return ss::json::json_return_type(ss::json::json_void());
-                  });
-            });
-      });
-}
-
-/*
- * Parse integer pairs from: ?target={\d,\d}* where each pair represent a
- * node-id and a shard-id, repsectively.
- */
-static std::vector<model::broker_shard>
-parse_target_broker_shards(const ss::sstring& param) {
-    std::vector<ss::sstring> parts;
-    boost::split(parts, param, boost::is_any_of(","));
-
-    if (parts.size() % 2 != 0) {
-        throw ss::httpd::bad_param_exception(
-          fmt::format("Invalid target parameter format: {}", param));
+    if (config::shard_local_cfg().enable_admin_api()) {
+        _admin.invoke_on_all(&admin_server::start).get0();
     }
-
-    std::vector<model::broker_shard> replicas;
-
-    for (auto i = 0u; i < parts.size(); i += 2) {
-        auto node = std::stoi(parts[i]);
-        auto shard = std::stoi(parts[i + 1]);
-
-        if (node < 0 || shard < 0) {
-            throw ss::httpd::bad_param_exception(
-              fmt::format("Invalid target {}:{}", node, shard));
-        }
-
-        replicas.push_back(model::broker_shard{
-          .node_id = model::node_id(node),
-          .shard = static_cast<uint32_t>(shard),
-        });
-    }
-
-    return replicas;
-}
-
-// TODO: factor out generic serialization from seastar http exceptions
-static security::scram_credential
-parse_scram_credential(const rapidjson::Document& doc) {
-    if (!doc.IsObject()) {
-        throw ss::httpd::bad_request_exception(fmt::format("Not an object"));
-    }
-
-    if (!doc.HasMember("algorithm") || !doc["algorithm"].IsString()) {
-        throw ss::httpd::bad_request_exception(
-          fmt::format("String algo missing"));
-    }
-    const auto algorithm = std::string_view(
-      doc["algorithm"].GetString(), doc["algorithm"].GetStringLength());
-
-    if (!doc.HasMember("password") || !doc["password"].IsString()) {
-        throw ss::httpd::bad_request_exception(
-          fmt::format("String password smissing"));
-    }
-    const auto password = doc["password"].GetString();
-
-    security::scram_credential credential;
-
-    if (algorithm == security::scram_sha256_authenticator::name) {
-        credential = security::scram_sha256::make_credentials(
-          password, security::scram_sha256::min_iterations);
-
-    } else if (algorithm == security::scram_sha512_authenticator::name) {
-        credential = security::scram_sha512::make_credentials(
-          password, security::scram_sha512::min_iterations);
-
-    } else {
-        throw ss::httpd::bad_request_exception(
-          fmt::format("Unknown scram algorithm: {}", algorithm));
-    }
-
-    return credential;
-}
-
-void application::admin_register_security_routes(ss::http_server& server) {
-    ss::httpd::security_json::create_user.set(
-      server._routes, [this](std::unique_ptr<ss::httpd::request> req) {
-          rapidjson::Document doc;
-          doc.Parse(req->content.data());
-
-          auto credential = parse_scram_credential(doc);
-
-          if (!doc.HasMember("username") || !doc["username"].IsString()) {
-              throw ss::httpd::bad_request_exception(
-                fmt::format("String username missing"));
-          }
-
-          auto username = security::credential_user(
-            doc["username"].GetString());
-
-          return controller->get_security_frontend()
-            .local()
-            .create_user(username, credential, model::timeout_clock::now() + 5s)
-            .then([this](std::error_code err) {
-                vlog(_log.debug, "Creating user {}:{}", err, err.message());
-                if (err) {
-                    throw ss::httpd::bad_request_exception(
-                      fmt::format("Creating user: {}", err.message()));
-                }
-                return ss::make_ready_future<ss::json::json_return_type>(
-                  ss::json::json_return_type(ss::json::json_void()));
-            });
-      });
-
-    ss::httpd::security_json::delete_user.set(
-      server._routes, [this](std::unique_ptr<ss::httpd::request> req) {
-          auto user = security::credential_user(
-            model::topic(req->param["user"]));
-
-          return controller->get_security_frontend()
-            .local()
-            .delete_user(user, model::timeout_clock::now() + 5s)
-            .then([this](std::error_code err) {
-                vlog(_log.debug, "Deleting user {}:{}", err, err.message());
-                if (err) {
-                    throw ss::httpd::bad_request_exception(
-                      fmt::format("Deleting user: {}", err.message()));
-                }
-                return ss::make_ready_future<ss::json::json_return_type>(
-                  ss::json::json_return_type(ss::json::json_void()));
-            });
-      });
-
-    ss::httpd::security_json::update_user.set(
-      server._routes, [this](std::unique_ptr<ss::httpd::request> req) {
-          auto user = security::credential_user(
-            model::topic(req->param["user"]));
-
-          rapidjson::Document doc;
-          doc.Parse(req->content.data());
-
-          auto credential = parse_scram_credential(doc);
-
-          return controller->get_security_frontend()
-            .local()
-            .update_user(user, credential, model::timeout_clock::now() + 5s)
-            .then([this](std::error_code err) {
-                vlog(_log.debug, "Updating user {}:{}", err, err.message());
-                if (err) {
-                    throw ss::httpd::bad_request_exception(
-                      fmt::format("Updating user: {}", err.message()));
-                }
-                return ss::make_ready_future<ss::json::json_return_type>(
-                  ss::json::json_return_type(ss::json::json_void()));
-            });
-      });
-
-    ss::httpd::security_json::list_users.set(
-      server._routes, [this](std::unique_ptr<ss::httpd::request>) {
-          std::vector<ss::sstring> users;
-          for (const auto& [user, _] :
-               controller->get_credential_store().local()) {
-              users.push_back(user());
-          }
-          return ss::make_ready_future<ss::json::json_return_type>(
-            std::move(users));
-      });
-}
-
-void application::admin_register_kafka_routes(ss::http_server& server) {
-    ss::httpd::kafka_json::kafka_transfer_leadership.set(
-      server._routes, [this](std::unique_ptr<ss::httpd::request> req) {
-          auto topic = model::topic(req->param["topic"]);
-
-          model::partition_id partition;
-          try {
-              partition = model::partition_id(
-                std::stoll(req->param["partition"]));
-          } catch (...) {
-              throw ss::httpd::bad_param_exception(fmt::format(
-                "Partition id must be an integer: {}",
-                req->param["partition"]));
-          }
-
-          if (partition() < 0) {
-              throw ss::httpd::bad_param_exception(
-                fmt::format("Invalid partition id {}", partition));
-          }
-
-          std::optional<model::node_id> target;
-          if (auto node = req->get_query_param("target"); !node.empty()) {
-              try {
-                  target = model::node_id(std::stoi(node));
-              } catch (...) {
-                  throw ss::httpd::bad_param_exception(
-                    fmt::format("Target node id must be an integer: {}", node));
-              }
-              if (*target < 0) {
-                  throw ss::httpd::bad_param_exception(
-                    fmt::format("Invalid target node id {}", *target));
-              }
-          }
-
-          vlog(
-            _log.info,
-            "Leadership transfer request for leader of topic-partition {}:{} "
-            "to node {}",
-            topic,
-            partition,
-            target);
-
-          model::ntp ntp(model::kafka_namespace, topic, partition);
-
-          auto shard = shard_table.local().shard_for(ntp);
-          if (!shard) {
-              throw ss::httpd::not_found_exception(fmt::format(
-                "Topic partition {}:{} not found", topic, partition));
-          }
-
-          return partition_manager.invoke_on(
-            *shard,
-            [ntp = std::move(ntp),
-             target](cluster::partition_manager& pm) mutable {
-                auto partition = pm.get(ntp);
-                if (!partition) {
-                    throw ss::httpd::not_found_exception();
-                }
-                return partition->transfer_leadership(target).then(
-                  [](std::error_code err) {
-                      if (err) {
-                          throw ss::httpd::server_error_exception(fmt::format(
-                            "Leadership transfer failed: {}", err.message()));
-                      }
-                      return ss::json::json_return_type(ss::json::json_void());
-                  });
-            });
-      });
-
-    ss::httpd::partition_json::kafka_move_partition.set(
-      server._routes, [this](std::unique_ptr<ss::httpd::request> req) {
-          auto topic = model::topic(req->param["topic"]);
-
-          model::partition_id partition;
-          try {
-              partition = model::partition_id(
-                std::stoll(req->param["partition"]));
-          } catch (...) {
-              throw ss::httpd::bad_param_exception(fmt::format(
-                "Partition id must be an integer: {}",
-                req->param["partition"]));
-          }
-
-          if (partition() < 0) {
-              throw ss::httpd::bad_param_exception(
-                fmt::format("Invalid partition id {}", partition));
-          }
-
-          std::optional<std::vector<model::broker_shard>> replicas;
-          if (auto node = req->get_query_param("target"); !node.empty()) {
-              try {
-                  replicas = parse_target_broker_shards(node);
-              } catch (...) {
-                  throw ss::httpd::bad_param_exception(fmt::format(
-                    "Invalid target format {}: {}",
-                    node,
-                    std::current_exception()));
-              }
-          }
-
-          // this can be removed when we have more sophisticated machinary in
-          // redpanda itself for automatically selecting target node/shard.
-          if (!replicas || replicas->empty()) {
-              throw ss::httpd::bad_request_exception(
-                "Partition movement requires target replica set");
-          }
-
-          model::ntp ntp(model::kafka_namespace, topic, partition);
-
-          vlog(
-            _log.debug,
-            "Request to change ntp {} replica set to {}",
-            ntp,
-            replicas);
-
-          return controller->get_topics_frontend()
-            .local()
-            .move_partition_replicas(
-              ntp, *replicas, model::timeout_clock::now() + 5s)
-            .then([this, ntp, replicas](std::error_code err) {
-                vlog(
-                  _log.debug,
-                  "Result changing ntp {} replica set to {}: {}:{}",
-                  ntp,
-                  replicas,
-                  err,
-                  err.message());
-                if (err) {
-                    throw ss::httpd::bad_request_exception(
-                      fmt::format("Error moving partition: {}", err.message()));
-                }
-                return ss::make_ready_future<ss::json::json_return_type>(
-                  ss::json::json_return_type(ss::json::json_void()));
-            });
-      });
 }
